refactor(architecture): the key adapter leaves the row, and the caret prop nobody set leaves the contract
The architecture-first-principles pass over this diff (C1–C7, kolu's skill),
two findings and their fixes.

**C2 · consumer-ergonomics — a knob nobody turns.** `MdeProps.caret` was
documented ("where the caret goes when this editor OPENS") and set by no
caller: the two gestures that have an opinion about where a caret opens are a
split and a merge, and both are a TITLE's. What actually needs it is the swap
between the editor's two faces, which is internal. So it leaves the public
contract and becomes `Placed`, the one thing the faces hand each other, and
the prop list is what a caller can answer.

**C3 · graduation / layer placement — a document was importing a ROW.**
`keyHandler` lived in `edit/RowEditor.tsx` because rows were its only callers;
the document editor made that a page reaching into a row's component module
for a function about keys. It graduates to `client/keying.ts` beside the
registry it asks: `keys.ts` decides what a key MEANS and is pure of the DOM,
`keying.ts` is the adapter that reads an element and spends the event. Three
surfaces ask it now — a title, a note, a document — and none of them imports
another's component to do it.

Nothing else moved. C1 found no hand-rolled duplicate (the debounce, the
serial queue, the chunk arrival and the storage circuit are all the
receptacles that already exist); C4's one-hop audit of what the new code
stands on is `arriving.ts`, `preference.ts`, `queue.ts` and `markdown/tags.ts`,
each used as documented; C6's five principles are answered in the headers —
notably P5, where `left()` is the guarantee relocated to the endpoint that
knows (the component doing the removing says so, rather than the app guessing
from `isConnected`).
