+++ packages/web/src/client/Tree.tsx
@@ -90,7 +90,8 @@ import { matchedAttr, useNarrowed } from "./filter/narrowed.tsx"
 import { onATag } from "./filter/tag.ts"
 import { useUndo } from "./edit/undoing.ts"
 import { NewRow } from "./edit/NewRow.tsx"
-import { DescEditor, keyHandler, Said, TitleEditor } from "./edit/RowEditor.tsx"
+import { DescEditor, Said, TitleEditor } from "./edit/RowEditor.tsx"
+import { keyHandler } from "./keying.ts"
 import { setFolded } from "./fold/memory.ts"
 import { createFoldReading } from "./fold/reading.ts"
 import { foldIdOf, foldOf, foldsUnder } from "./fold/rows.ts"
